Moved components into folders for organization.
- Created a Template folder and moved related Template components to the folder.
- Created an Edit folder and moved related Edit components to the folder.
