+++ src/components/Template/DivComp.js
