+++ src/components/Template/ImgComp.js
