+++ src/components/index.js
@@ -1,10 +1,10 @@
-export { default as DivComp } from './DivComp';
-export { default as PComp } from './PComp';
-export { default as ImgComp } from './ImgComp';
-export { default as ListComp } from './ListComp';
-export { default as ListItem } from './ListItem';
+export { default as DivComp } from './Template/DivComp';
+export { default as PComp } from './Template/PComp';
+export { default as ImgComp } from './Template/ImgComp';
+export { default as ListComp } from './Template/ListComp';
+export { default as ListItem } from './Template/ListItem';
 export { default as NavLink } from './NavLink';
-export { default as Header } from './Header';
-export { default as Footer } from './Footer';
-export { default as FontMenu } from './FontMenu';
-export { default as ColorMenu } from './ColorMenu';
+export { default as Header } from './Template/Header';
+export { default as Footer } from './Template/Footer';
+export { default as FontMenu } from './Edit/FontMenu';
+export { default as ColorMenu } from './Edit/ColorMenu';
